handle: add global status handler
