+++ statistics/handle/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//sessiontxn",
         "//statistics",
         "//statistics/handle/cache",
+        "//statistics/handle/globalstats",
         "//statistics/handle/lockstats",
         "//statistics/handle/metrics",
         "//table",
